backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
pytz==2025.2
PyYAML==6.0.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
